Abstract out Stub
