+++ src/main/java/seedu/address/logic/commands/RedoCommand.java
@@ -1,12 +1,11 @@
 package seedu.address.logic.commands;
 
+import static java.util.Objects.requireNonNull;
+
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.messages.RedoMessages;
-import seedu.address.logic.messages.UndoMessages;
 import seedu.address.model.Model;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * Redo Command.
