+++ src/main/java/seedu/address/logic/commands/UndoCommand.java
@@ -1,17 +1,11 @@
 package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.messages.AddMessages;
 import seedu.address.logic.messages.UndoMessages;
 import seedu.address.model.Model;
-import seedu.address.model.person.Person;
 
 /**
  * Undo Command.
