+++ src/main/java/seedu/address/storage/JsonAdaptedPerson.java
@@ -199,12 +199,12 @@ class JsonAdaptedPerson {
                     modelSkill, modelCommission);
         }
 
-//        if (note == null) {
-//            throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, Note.class.getSimpleName()));
-//        }
-//        if (!Note.isValidNote(note)) {
-//            throw new IllegalValueException(Note.MESSAGE_CONSTRAINTS);
-//        }
+        if (note == null) {
+            throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT, Note.class.getSimpleName()));
+        }
+        if (!Note.isValidNote(note)) {
+            throw new IllegalValueException(Note.MESSAGE_CONSTRAINTS);
+        }
         final Note modelNote = new Note(note);
 
         return new Person(modelName, modelPhone, modelEmail, modelAddress, modelNote, modelTags);
