+++ src/test/java/seedu/address/logic/stubs/ModelStubAcceptingPersonAdded.java
@@ -1,0 +1,33 @@
+package seedu.address.logic.stubs;
+
+import seedu.address.model.AddressBook;
+import seedu.address.model.ReadOnlyAddressBook;
+import seedu.address.model.person.Person;
+
+import java.util.ArrayList;
+
+import static java.util.Objects.requireNonNull;
+
+/**
+ * A Model stub that always accept the person being added.
+ */
+public class ModelStubAcceptingPersonAdded extends ModelStub {
+    public final ArrayList<Person> personsAdded = new ArrayList<>();
+
+    @Override
+    public boolean hasPerson(Person person) {
+        requireNonNull(person);
+        return personsAdded.stream().anyMatch(person::isSamePerson);
+    }
+
+    @Override
+    public void addPerson(Person person) {
+        requireNonNull(person);
+        personsAdded.add(person);
+    }
+
+    @Override
+    public ReadOnlyAddressBook getAddressBook() {
+        return new AddressBook();
+    }
+}
