+++ src/test/java/seedu/address/logic/stubs/ModelStubWithPerson.java
@@ -1,0 +1,23 @@
+package seedu.address.logic.stubs;
+
+import seedu.address.model.person.Person;
+
+import static java.util.Objects.requireNonNull;
+
+/**
+ * A Model stub that contains a single person.
+ */
+public class ModelStubWithPerson extends ModelStub {
+    private final Person person;
+
+    public ModelStubWithPerson(Person person) {
+        requireNonNull(person);
+        this.person = person;
+    }
+
+    @Override
+    public boolean hasPerson(Person person) {
+        requireNonNull(person);
+        return this.person.isSamePerson(person);
+    }
+}
